refactor: move exit-status mapping out of main
Adding the interrupt code pushed main.rs to 507 code lines, past the
500-line limit. The exit-status handling is the natural piece to lift out:
it is a closed responsibility with conventions of its own to hold —
docker's 126/127 for a command that cannot be launched, 130 for an
interrupt — and nothing else in main needs to know them.

Pure move. Same 67 binary tests pass before and after.

Part of #1080
