+++ internal/main.rs
@@ -11,11 +11,13 @@ use clap::CommandFactory;
 mod autostart_cmd;
 mod cli;
 mod dispatch;
+mod exit_status;
 mod generate;
 mod resolve;
 mod startup;
 
 use cli::*;
+use exit_status::{command_failure_exit_code, interrupt_exit_code, print_error};
 use generate::write_quadlet;
 use resolve::*;
 use startup::{init_tracing, internal_error_notice, is_label_only, is_mutating, parse_cli};
@@ -128,33 +130,6 @@ fn print_command_help(commands: &[String]) -> podup::Result<()> {
 	Ok(())
 }
 
-/// Map a failed launch onto docker's conventional exit codes by inspecting the
-/// OCI/crun error text: a "command not found" failure → 127, a
-/// "not executable"/"permission denied"/"exec format error" failure → 126,
-/// anything else → 1. Pure string inspection so it is unit-testable.
-/// Exit status for an attached `up` ended by a signal: 128 + SIGINT, the shell
-/// convention, and what `docker compose up` returns for SIGTERM as well.
-const fn interrupt_exit_code() -> i32 {
-	130
-}
-
-fn command_failure_exit_code(msg: &str) -> i32 {
-	let m = msg.to_ascii_lowercase();
-	let not_found = m.contains("executable file not found")
-		|| m.contains("not found in $path")
-		|| (m.contains("oci runtime") && m.contains("no such file"));
-	let not_executable = m.contains("exec format error")
-		|| m.contains("not executable")
-		|| (m.contains("oci runtime") && m.contains("permission denied"));
-	if not_found {
-		127
-	} else if not_executable {
-		126
-	} else {
-		1
-	}
-}
-
 /// Whether a `down` invocation should tear the project down purely by its
 /// `podup.project` label: the command is `down`, an explicit project name was
 /// given (`-p` / `COMPOSE_PROJECT_NAME`), and no compose file resolves on disk.
@@ -162,20 +137,6 @@ fn command_failure_exit_code(msg: &str) -> i32 {
 /// file is absent. Pure so it is unit-testable without a Podman daemon.
 fn down_by_label_path(command: &Commands, project: Option<&str>, compose_present: bool) -> bool {
 	matches!(command, Commands::Down { .. }) && project.is_some() && !compose_present
-}
-
-/// Print a top-level error to stderr with a colour-aware bold-red `error:` label.
-/// anstream strips the styling when stderr is not a terminal or colour is off.
-fn print_error(e: &podup::ComposeError) {
-	use std::io::Write;
-	let style = podup::ui::error_style();
-	let mut err = anstream::stderr();
-	let _ = writeln!(
-		err,
-		"podup: {}error:{} {e}",
-		style.render(),
-		style.render_reset()
-	);
 }
 
 /// Orchestrate one CLI invocation: parse args, then short-circuit the commands
@@ -630,41 +591,6 @@ mod down_by_label_tests {
 		// Only `down` is routed by label here; another command with `-p` and no file
 		// must not be diverted.
 		assert!(!down_by_label_path(&Commands::Watch, Some("proj"), false));
-	}
-}
-
-#[cfg(test)]
-mod exit_code_tests {
-	use super::command_failure_exit_code;
-
-	#[test]
-	fn not_found_maps_to_127() {
-		assert_eq!(
-			command_failure_exit_code(
-				"podman error: crun: executable file `foo` not found in $PATH: \
-				 No such file or directory: OCI runtime command not found error"
-			),
-			127
-		);
-		assert_eq!(
-			command_failure_exit_code("OCI runtime error: ...: no such file or directory"),
-			127
-		);
-	}
-
-	#[test]
-	fn not_executable_maps_to_126() {
-		assert_eq!(
-			command_failure_exit_code("OCI runtime error: permission denied"),
-			126
-		);
-		assert_eq!(command_failure_exit_code("exec format error"), 126);
-	}
-
-	#[test]
-	fn unrelated_errors_map_to_1() {
-		assert_eq!(command_failure_exit_code("some other failure"), 1);
-		assert_eq!(command_failure_exit_code("container is restarting"), 1);
 	}
 }
